Allow certain methods to be ignored. This is useful in situations where the reflection 'finds' methods that are not practically present.
For example, an interface with a method using generics, will create a ghost method which
is overwritten in a typed version, and itself is not really present:

public interface IGenerator<T extends IChemObject>  {
  public IRenderingElement generate(T object, RendererModel model);
}

The coverage test will now 'reflect' a method generate(IChemObject, RendererModel)
which will never be implemented, becuase the implementing class witll implement
this method with whatever T it is written for (e.g. IAtomContainer).

Now, I do not know how to get the coverage test to 'reflect' on the ghost nature
of this method, so a more general ignore list is implemented.